todo를 업데이트하는 로직을 클래스로 분리하라
