+++ app-server/src/test/java/com/growth/task/todo/application/TodoServiceTest.java
@@ -4,11 +4,7 @@ import com.growth.task.task.domain.Tasks;
 import com.growth.task.task.repository.TasksRepository;
 import com.growth.task.todo.domain.Todos;
 import com.growth.task.todo.dto.request.TodoAddRequest;
-import com.growth.task.todo.dto.request.TodoUpdateRequest;
-import com.growth.task.todo.dto.response.TodoUpdateResponse;
-import com.growth.task.todo.enums.Status;
 import com.growth.task.todo.exception.TaskNotFoundException;
-import com.growth.task.todo.exception.TodoNotFoundException;
 import com.growth.task.todo.repository.TodosRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -20,10 +16,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class TodoServiceTest {
@@ -38,7 +35,6 @@ class TodoServiceTest {
     private final Long TASK_ID1 = 1L;
     private final Long TASK_ID2 = 2L;
     private final String WHAT_TO_DO = "디자인패턴의 아름다움 스터디";
-    private final String NEW_WHAT_TO_DO = "쿠버네티스 입문";
 
     @BeforeEach
     void setUp() {
@@ -97,97 +93,4 @@ class TodoServiceTest {
         }
     }
 
-    @Nested
-    @DisplayName("TodoService 의 update 메서드는")
-    class Describe_update {
-        private final Long TODO_ID1 = 1L;
-
-        @Nested
-        @DisplayName("todoId 로 todos 가 확인되는 경우")
-        class Context_whenTodoIdExists {
-            private final TodoUpdateRequest todoUpdateRequest_todo = TodoUpdateRequest.builder()
-                    .todo(NEW_WHAT_TO_DO)
-                    .build();
-
-            private final TodoUpdateRequest todoUpdateRequest_status = TodoUpdateRequest.builder()
-                    .status(Status.DONE)
-                    .build();
-
-            private final TodoUpdateRequest todoUpdateRequest_todo_and_status = TodoUpdateRequest.builder()
-                    .todo(NEW_WHAT_TO_DO)
-                    .status(Status.DONE)
-                    .build();
-
-            private final Tasks tasks = Tasks.builder()
-                    .taskId(TASK_ID1)
-                    .build();
-
-            private final Todos todos = Todos.builder()
-                    .todoId(TODO_ID1)
-                    .task(tasks)
-                    .todo(WHAT_TO_DO)
-                    .status(Status.PROGRESS)
-                    .build();
-
-            @BeforeEach
-            void setUp() {
-                given(todosRepository.findById(TODO_ID1)).willReturn(Optional.of(todos));
-            }
-
-            @Test
-            @DisplayName("todo 에 대한 TodoUpdateRequest 가 저장된다.")
-            void It_updateTheTodo() {
-                Todos todos = todoService.update(TODO_ID1, todoUpdateRequest_todo);
-                TodoUpdateResponse response = new TodoUpdateResponse(TODO_ID1, todos);
-
-                assertAll(
-                        () -> assertThat(response.getTodo()).isEqualTo(NEW_WHAT_TO_DO),
-                        () -> assertThat(response.getStatus()).isEqualTo(Status.PROGRESS)
-                );
-            }
-
-            @Test
-            @DisplayName("status 에 대한 TodoUpdateRequest 가 저장된다.")
-            void It_updateTheStatus() {
-                Todos todos = todoService.update(TODO_ID1, todoUpdateRequest_status);
-                TodoUpdateResponse response = new TodoUpdateResponse(TODO_ID1, todos);
-
-                assertAll(
-                        () -> assertThat(response.getTodo()).isEqualTo(WHAT_TO_DO),
-                        () -> assertThat(response.getStatus()).isEqualTo(Status.DONE)
-                );
-            }
-
-            @Test
-            @DisplayName("todo 와 status 에 대한 TodoUpdateRequest 가 저장된다.")
-            void It_updateTheTodoAndStatus() {
-                Todos todos = todoService.update(TODO_ID1, todoUpdateRequest_todo_and_status);
-                TodoUpdateResponse response = new TodoUpdateResponse(TODO_ID1, todos);
-
-                assertAll(
-                        () -> assertThat(response.getTodo()).isEqualTo(NEW_WHAT_TO_DO),
-                        () -> assertThat(response.getStatus()).isEqualTo(Status.DONE)
-                );
-            }
-        }
-
-        @Nested
-        @DisplayName("todoId 로 Todos 가 확인되지 않는 경우")
-        class Context_whenTodoIdDoesNotExist {
-            private final TodoUpdateRequest todoUpdateRequest = TodoUpdateRequest.builder()
-                    .todo(NEW_WHAT_TO_DO)
-                    .build();
-
-            @BeforeEach
-            void setUp() {
-                when(todosRepository.findById(TODO_ID1)).thenReturn(Optional.empty());
-            }
-
-            @Test
-            @DisplayName("TodoNotFoundException 오류를 던진다.")
-            void It_throws_TodoNotFoundException() {
-                assertThrows(TodoNotFoundException.class, () -> todoService.update(TODO_ID1, todoUpdateRequest));
-            }
-        }
-    }
 }
